requirements.txt
# Core
python-dotenv==1.0.0
PyYAML==6.0.1

# Apache Airflow
apache-airflow==2.7.3
apache-airflow-providers-mongo==3.4.0
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-amazon==8.8.0
apache-airflow-providers-apache-spark==4.3.0

# Data Processing
pyspark==3.5.0
pandas==2.1.4
numpy==1.26.2

# NLP / ML
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
nltk==3.8.1

# Databases
pymongo==4.6.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AWS
boto3==1.34.11
botocore==1.34.11

# Web API
Flask==3.0.0
Flask-RESTful==0.3.10
flask-cors==4.0.0

# Utilities
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
